HW12/main.py
import requests
from fastapi import FastAPI
import databases
import sqlalchemy
from pydantic import BaseModel

# Инициализация FastAPI приложения
app = FastAPI()

# Подключение к базе данных SQLite
DATABASE_URL = "sqlite:///./clinic.db"
database = databases.Database(DATABASE_URL)
metadata = sqlalchemy.MetaData()

# Определение таблиц в базе данных: Клиенты, Консультации
clients = sqlalchemy.Table(
    "Clients",
    metadata,
    sqlalchemy.Column("ClientId", sqlalchemy.Integer, primary_key=True),
    sqlalchemy.Column("Document", sqlalchemy.String(50)),
    sqlalchemy.Column("SurName", sqlalchemy.String(50)),
    sqlalchemy.Column("FirstName", sqlalchemy.String(50)),
    sqlalchemy.Column("Birthday", sqlalchemy.String(50)),
)

# Создание базы данных
engine = sqlalchemy.create_engine(
    DATABASE_URL, connect_args={"check_same_thread": False}
)
metadata.create_all(engine)


# Модели данных Pydantic для клиентов и консультаций
class Client(BaseModel):
    Document: str
    SurName: str
    FirstName: str
    Birthday: str


# Определение эндпоинтов FastAPI
@app.get("/data")
def get_data():
    # Реализация логики для получения данных из базы данных или внешнего сервиса
    return {"data": {"detail": "Not Found"}}


@app.post("/data")
def add_data():
    # Реализация логики для добавления данных в базу данных или внешний сервис
    return {"added_data": {"detail": "Not Found"}}


@app.put("/data/{data_id}")
def update_data(data_id: int):
    # Реализация логики для обновления данных в базе данных или внешнем сервисе
    return {"updated": {"detail": "Not Found"}}


@app.delete("/data/{data_id}")
def delete_data(data_id: int):
    # Реализация логики для удаления данных в базе данных или внешнем сервисе
    return {"deleted": False}


@app.get("/get_data")
def get_data():
    return {"data": "some data"}


@app.post("/add_data")
def add_data():
    return {"added_data": "data added"}


@app.put("/update_data/{data_id}")
def update_data(data_id: int):
    return {"updated": f"данные с ID {data_id} обновлены"}


@app.delete("/delete_data/{data_id}")
def delete_data(data_id: int):
    return {"deleted": f"данные с ID {data_id} удалены"}


@app.get("/healthcheck")
def healthcheck():
    return {"status": "OK"}


@app.get("/items/{item_id}")
def read_item(item_id: int, q: str = None):
    return {"item_id": item_id, "q": q}
